[12]: added income source creating, adding income to account

+++ src/main/java/ru/kpfu/itis/jblab/controller/OperationController.java
@@ -6,9 +6,11 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import ru.kpfu.itis.jblab.model.Account;
 import ru.kpfu.itis.jblab.model.Expense;
+import ru.kpfu.itis.jblab.model.Income;
 import ru.kpfu.itis.jblab.model.Operation;
 import ru.kpfu.itis.jblab.service.AccountService;
 import ru.kpfu.itis.jblab.service.ExpenseService;
+import ru.kpfu.itis.jblab.service.IncomeService;
 import ru.kpfu.itis.jblab.service.OperationService;
 
 /**
@@ -19,12 +21,14 @@ public class OperationController {
     private final OperationService operationService;
     private final ExpenseService expenseService;
     private final AccountService accountService;
+    private final IncomeService incomeService;
 
     @Autowired
-    public OperationController(OperationService operationService, ExpenseService expenseService, AccountService accountService) {
+    public OperationController(OperationService operationService, ExpenseService expenseService, AccountService accountService, IncomeService incomeService) {
         this.operationService = operationService;
         this.expenseService = expenseService;
         this.accountService = accountService;
+        this.incomeService = incomeService;
     }
 
     @RequestMapping(value = "/operation/{id}/delete")
@@ -38,7 +42,15 @@ public class OperationController {
                 Expense expense = expenseService.getByOperationId(operationId);
                 account.setBalance(account.getBalance() + operation.getAmount());
                 accountService.update(account);
-                expenseService.delete(expense);
+                if (expense != null)
+                    expenseService.delete(expense);
+                break;
+            case "income":
+                Income income = incomeService.getByOperationId(operationId);
+                account.setBalance(account.getBalance() - operation.getAmount());
+                accountService.update(account);
+                if (income != null)
+                    incomeService.delete(income);
                 break;
         }
         operationService.delete(operation);
